% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci2num.R
\name{ci2num}
\alias{ci2num}
\title{Estimate binomial numerator and denominator given an estimate and uncertainty interval for a proportion}
\usage{
ci2num(est, lower, upper)
}
\arguments{
\item{est}{Point estimate}

\item{lower}{Lower credible limit}

\item{upper}{Upper credible limit}
}
\description{
Based on least squares method provided by the SHELF package TODO REF
}
\examples{
est <- 3.00 / 100 
upper <- 3.52 / 100 
lower <- 2.60 / 100 
ci2num(est, lower, upper)

}
