% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{disbayes}
\alias{disbayes}
\title{Bayesian estimation of chronic disease epidemiology}
\usage{
disbayes(data, inc_num = NULL, inc_denom = NULL, inc = NULL,
  inc_lower = NULL, inc_upper = NULL, prev_num = NULL,
  prev_denom = NULL, prev = NULL, prev_lower = NULL,
  prev_upper = NULL, mort_num = NULL, mort_denom = NULL,
  mort = NULL, mort_lower = NULL, mort_upper = NULL, age = "age",
  smooth = TRUE, cf_init = 0.01, eqage = 30, sprior = 1)
}
\arguments{
\item{data}{Data frame containing some of the variables below.  The variables below are provided as character strings naming columns in this data frame.   For incidence and mortality, one of the three combinations of variables must be specified

(1) numerator and denominator
(2) estimate and denominator
(3) estimate with lower and upper credible limits

For estimates based on registry data that cover the whole population, then the denominator will be the population size.}

\item{inc_num}{Numerator for the incidence data, assumed to represent the observed number of new cases within a year among a population of size \code{inc_denom}.}

\item{inc_denom}{Denominator for the incidence data.

The function \code{\link{ci2num}} can be used to convert a published estimate and interval for a proportion to an implicit numerator and denominator.

Note that to include extra uncertainty beyond that implied by a published interval, the numerator and denominator could be multiplied by a constant, for example, multiplying both the numerator and denominator by 0.5 would give the data source half its original weight.}

\item{inc}{Estimate of incidence.}

\item{inc_lower}{Lower credible limit for the incidence estimate}

\item{inc_upper}{Upper credible limit for the incidence estimate}

\item{prev_num}{Numerator for the estimate of prevalence}

\item{prev_denom}{Numerator for the estimate of prevalence}

\item{prev}{Estimate of prevalence}

\item{prev_lower}{Lower credible limit for the prevalence estimate}

\item{prev_upper}{Upper credible limit for the prevalence estimate}

\item{mort_num}{Numerator for the estimate of the mortality rate}

\item{mort_denom}{Denominator for the estimate of the mortality rate (e.g the population size, if )}

\item{mort}{Estimate of the mortality rate}

\item{mort_lower}{Lower credible limit for the mortality estimate}

\item{mort_upper}{Upper credible limit for the mortality estimate}

\item{age}{Variable in the data indicating the year of age}

\item{smooth}{Case fatality is modelled as a smooth function of age, using a spline.  If \code{smooth=FALSE} then a priori the case fatalities are assumed to be independent for each year of age.

The unsmoothed model is useful for determining how much information is in the data. That is, if the posterior from this model is identical to the prior for a certain age, then there is no information in the data alone about case fatality at that age, indicating that some other structural assumption (such as a smooth function of age) or external data are equired to give more precise estimates.}

\item{cf_init}{Initial guess at a typical case fatality value, for an average age.}

\item{eqage}{Case fatality is assumed to be equal for all ages below this age}

\item{sprior}{Rate of the exponential prior distribution used to penalise the coefficients of the spline model.}
}
\description{
Estimates a three-state disease model given data on incidence, prevalence and mortality.
}
\details{
Currently it is designed to estimate case fatality given data on at least mortality and incidence.

It could be modified to estimate incidence given prevalence and mortality (this would need a new Stan template model with incidence as a smooth function of age)

Intended to extend to hierarchical models - will test in context of METAHIT project

Other extensions to be motivated by applications
}
