% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes-package.R
\docType{data}
\name{ihdbristol}
\alias{ihdbristol}
\title{Ischaemic heart disease in Bristol}
\format{TODO}
\source{
Global Burden of Disease study
}
\usage{
ihdbristol
}
\description{
Ischaemic heart disease in Bristol
}
\keyword{datasets}
