% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes-package.R
\docType{data}
\name{ihdlondon}
\alias{ihdlondon}
\title{Ischaemic heart disease in London}
\format{TODO}
\source{
Global Burden of Disease study
}
\usage{
ihdlondon
}
\description{
Ischaemic heart disease in London
}
\keyword{datasets}
