% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.disbayes}
\alias{loo.disbayes}
\title{Leave-one-out cross validation for disbayes models}
\usage{
\method{loo}{disbayes}(x, outcome = "overall", ...)
}
\arguments{
\item{x}{A model fitted by \code{\link{disbayes}}.  Any of the computation methods
are supported.}

\item{outcome}{Either \code{"overall"}, to assess the fit to all data, or 
one of \code{"inc"}, \code{"prev"}, \code{"mort"} or \code{"rem"}, to assess the fit
to the incidence data, prevalence data, mortalidy data or remission data, respectively.}

\item{...}{Other arguments (currently unused).}
}
\value{
An object of class \code{"loo"} as defined by the \pkg{loo} package.
}
\description{
Leave-one-out cross validation for disbayes models
}
\seealso{
\code{\link{loo_indiv}} to return tidied observation-specific contributions
to the overall model fit computed here.
}
