% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{looi_disbayes}
\alias{looi_disbayes}
\title{Observation-level leave-one-out cross-validatory statistics from a disbayes fit}
\usage{
looi_disbayes(x)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}.}
}
\description{
Return observation-level leave-one-out cross-validatory statistics from a disbayes fit
as a tidy data frame.
}
\details{
The data frame has one row per observed age-specific mortality, incidence, prevalence and/or
remission data-point, containing leave-one-out cross validation statistics representing how
well the model would predict that observation if it were left out of the fit. 

These are computed with the \pkg{loo} package.
}
