% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{looi_disbayes_hier_gender}
\alias{looi_disbayes_hier_gender}
\title{Observation-level leave-one-out cross-validatory statistics from a disbayes_hier_gender fit}
\usage{
looi_disbayes_hier_gender(x)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes_hier}} with an additive gender effect, so that
the object has class \code{"disbayes_hier_gender"}}
}
\description{
Return observation-level leave-one-out cross-validatory statistics from a disbayes_hier_gender fit
as a tidy data frame.
}
