% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{plot.disbayes_hier}
\alias{plot.disbayes_hier}
\title{Quick plot of estimates from hierarchical disbayes models against age}
\usage{
\method{plot}{disbayes_hier}(x, variable = "cf", ci = FALSE, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes_hier}}}

\item{variable}{Name of the variable of interest to plot against age, by default case fatality rates.}

\item{ci}{Show 95\% credible intervals with ribbons.}

\item{...}{Other arguments. Currently unused}
}
\value{
A \code{ggplot2} object that can be printed to show the plot, or
customised by adding \code{geom}s.

Better plots can be drawn by \code{tidy}ing the object returned by \code{disbayes}, and using \code{ggplot2} directly on the tidy data frame that this produces.  See the vignette for examples.
}
\description{
Posterior medians and 95\% credible intervals for a quantity of interest are plotted against year of age.
}
