% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{plotfit_data_disbayes}
\alias{plotfit_data_disbayes}
\title{Create tidy data for a check of observed against fitted outcome probability estimates
from disbayes}
\usage{
plotfit_data_disbayes(x)
}
\arguments{
\item{x}{Fitted model from \code{\link{disbayes}}}
}
\value{
A data frame containing observed data in the form of outcome probabilities, as extracted by \code{\link{tidy_obsdat}}, and estimates of the corresponding probability parameters from the fitted model.
}
\description{
Create tidy data for a check of observed against fitted outcome probability estimates
from disbayes
}
