% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{plotfit_disbayes}
\alias{plotfit_disbayes}
\title{Graphical check of observed against fitted outcome probabilities from disbayes}
\usage{
plotfit_disbayes(x, agemin = 50)
}
\arguments{
\item{x}{Fitted model from \code{\link{disbayes}}}

\item{agemin}{Minimum age to show on the horizontal axis.}
}
\description{
The data behind the plot can be produced using \code{\link{plotfit_data_disbayes}},
to enable customised plots to be produced by hand with \code{ggplot2}.
}
