% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{summary.disbayes}
\alias{summary.disbayes}
\title{Summarise estimates from the model}
\usage{
\method{summary}{disbayes}(object, vars = NULL, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{disbayes}}}

\item{vars}{Names of the variables of interest to return.  If not supplied, all estimated quantities are returned.}

\item{...}{Other arguments. Currently unused}
}
\description{
Summarise estimates from the model
}
