% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{summary_disbayes_fit}
\alias{summary_disbayes_fit}
\title{Summarise estimates from the model  TODO DOC}
\usage{
summary_disbayes_fit(fit, vars = NULL)
}
\arguments{
\item{fit}{A Stan fitted model object}

\item{vars}{Names of variables indexed by age, e.g. \code{vars="cf"} if you want the case fatality estimates for all ages.   TODO describe other variables that can be extracted}
}
\description{
Summarise estimates from the model  TODO DOC
}
