% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{tidy.disbayes}
\alias{tidy.disbayes}
\title{Form a tidy data frame from the estimates from a disbayes fit}
\usage{
\method{tidy}{disbayes}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}}

\item{...}{Other arguments (currently unused)}
}
\value{
A data frame with one row per model parameter, giving summary statistics
for the posterior distribution for that parameter.   For array parameters, e.g. those
that depend on age or area, then the age and area are returned in separate columns,
to make it easier to summarise and plot the results, e.g. using \pkg{ggplot2}.
}
\description{
Simply call this after fitting disbayes, as, e.g.
```
res <- disbayes(...)
tidy(res)
```
}
