% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{tidy.disbayes_hier}
\alias{tidy.disbayes_hier}
\title{Return summary statistics from a hierarchical disbayes model}
\usage{
\method{tidy}{disbayes_hier}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes_hier}}}

\item{...}{Other arguments (currently unused)}
}
\description{
Return summary statistics from a hierarchical disbayes model as a tidy data frame.
}
\seealso{
\code{\link{tidy.disbayes}}
}
