% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{tidy.disbayes_hier_gender}
\alias{tidy.disbayes_hier_gender}
\title{Form a tidy data frame from the estimates from a disbayes_hier fit with an additive gender effect}
\usage{
\method{tidy}{disbayes_hier_gender}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes_hier}} with an additive gender effect}

\item{...}{Other arguments (currently unused)}
}
\description{
Form a tidy data frame from the estimates from a disbayes_hier fit with an additive gender effect
}
\seealso{
\code{\link{tidy.disbayes_hier}}, \code{\link{tidy.disbayes}}
}
