% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes-package.R
\docType{package}
\name{disbayes-package}
\alias{disbayes-package}
\title{The 'disbayes' package.}
\description{
Bayesian evidence synthesis for chronic disease epidemiology
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
