% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{disbayes_hier}
\alias{disbayes_hier}
\title{Bayesian estimation of chronic disease epidemiology from incomplete data - hierarchical model for case fatalities 
This is computationally intensive.}
\usage{
disbayes_hier(
  data,
  group,
  gender = NULL,
  inc_num = NULL,
  inc_denom = NULL,
  inc = NULL,
  inc_lower = NULL,
  inc_upper = NULL,
  prev_num = NULL,
  prev_denom = NULL,
  prev = NULL,
  prev_lower = NULL,
  prev_upper = NULL,
  mort_num = NULL,
  mort_denom = NULL,
  mort = NULL,
  mort_lower = NULL,
  mort_upper = NULL,
  rem_num = NULL,
  rem_denom = NULL,
  rem = NULL,
  rem_lower = NULL,
  rem_upper = NULL,
  age = "age",
  remission = FALSE,
  cf_init = 0.01,
  eqage = 30,
  model = "default",
  loo = TRUE,
  sprior = 1,
  nfold_int_guess = 5,
  nfold_int_upper = 100,
  nfold_slope_guess = 5,
  nfold_slope_upper = 100,
  mean_int_prior = c(0, 10),
  mean_slope_prior = c(5, 5),
  gender_int_priorsd = 0.82,
  gender_slope_priorsd = 0.82,
  iter = 10000,
  iter_train = 1000,
  s_opts = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame containing some of the variables below.  The
  variables below are provided as character strings naming columns in this
  data frame.   For incidence and mortality, one of the following three
  combinations of variables must be specified:

  (1) numerator and denominator (2) estimate and denominator (3) estimate
  with lower and upper credible limits

  For estimates based on registry data assumed to cover the whole population, then
  the denominator will be the population size.}

\item{group}{Variable in the data representing the area (or other grouping factor).}

\item{gender}{Variable in the data representing gender, or a binary grouping factor, that is treated as a fixed additive effect, so the linear effect of gender on log case fatality is the same in each area.   The data should then have one row per year of age and gender.  If \code{NULL} (the default) then the data are one homogenous gender, and there should be one row per year of age.}

\item{inc_num}{Numerator for the incidence data, assumed to represent the
observed number of new cases within a year among a population of size
\code{inc_denom}.}

\item{inc_denom}{Denominator for the incidence data.

  The function \code{\link{ci2num}} can be used to convert a published
  estimate and interval for a proportion to an implicit numerator and
  denominator.

  Note that to include extra uncertainty beyond that implied by a published
  interval, the numerator and denominator could be multiplied by a constant,
  for example, multiplying both the numerator and denominator by 0.5 would
  give the data source half its original weight.}

\item{inc}{Estimate of incidence.}

\item{inc_lower}{Lower credible limit for the incidence estimate}

\item{inc_upper}{Upper credible limit for the incidence estimate}

\item{prev_num}{Numerator for the estimate of prevalence}

\item{prev_denom}{Denominator for the estimate of prevalence (e.g. the size
of the survey used to obtain the prevalence estimate)}

\item{prev}{Estimate of prevalence}

\item{prev_lower}{Lower credible limit for the prevalence estimate}

\item{prev_upper}{Upper credible limit for the prevalence estimate}

\item{mort_num}{Numerator for the estimate of the mortality rate}

\item{mort_denom}{Denominator for the estimate of the mortality rate (e.g.
the population size, if the estimates were obtained from a comprehensive
register)}

\item{mort}{Estimate of the mortality rate}

\item{mort_lower}{Lower credible limit for the mortality estimate}

\item{mort_upper}{Upper credible limit for the mortality estimate}

\item{rem_num}{Numerator for the estimate of the remission rate}

\item{rem_denom}{Denominator for the estimate of the remission rate}

\item{rem}{Estimate of the remission rate}

\item{rem_lower}{Lower credible limit for the remission estimate}

\item{rem_upper}{Upper credible limit for the remission estimate}

\item{age}{Variable in the data indicating the year of age}

\item{remission}{Is remission from the disease permitted in the model
\code{TRUE} or \code{FALSE}.  If \code{TRUE}, then data must be provided, as
for incidence, mortality and prevalence.}

\item{cf_init}{Initial guess at a typical case fatality value, for an
average age.}

\item{eqage}{Case fatality and incidence are assumed to be equal for all ages
below this age.}

\item{model}{Three alternative models for case fatality are supported:

\code{"default"} (the default). Random intercepts and slopes.

\code{"interceptonly"}.  Random intercepts, but common slopes.

\code{"increasing"}. Case fatality is assumed to be an increasing function of age (note it is constant below \code{"eqage"} in all models).}

\item{loo}{Compute leave-one-out cross validation statistics.}

\item{sprior}{Rate of the exponential prior distribution used to penalise
the coefficients of the spline model.  The default of 1 should adapt
appropriately to the data, but higher values give stronger smoothing,
or lower values give weaker smoothing,  if required.}

\item{nfold_int_guess}{Prior guess at the ratio of case fatality between a high risk (97.5\% quantile) and low risk (2.5\% quantile) area.}

\item{nfold_int_upper}{Prior upper 95\% credible limit for the ratio in average case fatality between a high risk (97.5\% quantile) and low risk (2.5\% quantile) area.}

\item{nfold_slope_guess, nfold_slope_upper}{This argument and the next argument define the prior distribution for the variance in the random linear effects of age on log case fatality.   They define a prior guess and upper 95\% credible limit for the ratio of case fatality slopes between a high trend (97.5\% quantile) and low risk (2.5\% quantile) area.  (Note that the model is not exactly linear, since departures from linearity are defined through a spline model.  See the Jackson et al. paper for details.).}

\item{iter}{Number of MCMC iterations to use for the smoothed model fit.}

\item{iter_train}{Number of MCMC iterations to use for the unsmoothed (training)
model fit.}

\item{s_opts}{List of arguments to supply to the function \code{\link[mgcv]{s}}
for constructing a spline basis, e.g. \code{list(bs="cr")} to switch from the
default "thin plate" spline to a cubic spline.  Currently not implemented!}

\item{...}{Further arguments passed to \pkg{rstan}{sampling} to control
running of Stan, that are applied to both the "training" (unsmoothed) model
fit and the smoothed model fit.}
}
\description{
Bayesian estimation of chronic disease epidemiology from incomplete data - hierarchical model for case fatalities 
This is computationally intensive.
}
\details{
For full details see Jackson et al. REF.

Note that an independent incidence is estimated for each area and age group.  In principle this function
would be straightforward to extend to model incidences as a smooth function of age and hierarchically by area.
Please contact the author if you would be interested in this feature.
}
