% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes-package.R
\docType{data}
\name{ihdtrends}
\alias{ihdtrends}
\title{Trends in ischemic heart disease in England}
\format{
A data frame with columns:

\code{gender}: \code{"male"} or \code{"female"}. 

\code{age}: Year of age. 

\code{year}: Calendar year.

\code{p2017}: Estimated ratio between the outcome in the calendar
year and the outcome in 2017.

\code{outcome}: Outcome referred to (incidence or case fatality).
}
\source{
TODO
}
\usage{
ihdtrends
}
\description{
Trends in ischemic heart disease in England
}
\details{
The data were processed to

The script used to do this is available from REF GITHUB TODO
}
\references{
TODO our paper when available
}
\keyword{datasets}
