% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{looi_disbayes}
\alias{looi_disbayes}
\title{Observation-level leave-one-out cross-validatory statistics from a disbayes fit}
\usage{
looi_disbayes(x, looname = "loo")
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}}

\item{looname}{\code{loo} for smoothed model, \code{loou} for unsmoothed model.}
}
\description{
Return observation-level leave-one-out cross-validatory statistics from a disbayes fit
as a tidy data frame.
}
