% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes_hier.R
\name{looi_disbayes_hier}
\alias{looi_disbayes_hier}
\title{Observation-level leave-one-out cross-validatory statistics from a disbayes_hier fit}
\usage{
looi_disbayes_hier(x)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes_hier}}}
}
\description{
Return observation-level leave-one-out cross-validatory statistics from a disbayes_hier_gender fit
as a tidy data frame.
}
