% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{plot.disbayes}
\alias{plot.disbayes}
\title{Plot estimates from the model against age}
\usage{
\method{plot}{disbayes}(x, variable = "cf", ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}}

\item{variable}{Name of the variable of interest to plot against age}

\item{...}{Other arguments. Currently unused}
}
\description{
Posterior medians and 95% credible intervals for a quantity of interest are plotted against year of age.
}
