% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{tidy.disbayes}
\alias{tidy.disbayes}
\title{Form a tidy data frame from the estimates from a disbayes fit}
\usage{
\method{tidy}{disbayes}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}}

\item{...}{Other arguments (currently unused)}
}
\description{
Form a tidy data frame from the estimates from a disbayes fit
}
