% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disbayes.R
\name{tidy.disbayes_trend}
\alias{tidy.disbayes_trend}
\title{Tidy output for a disbayes model with a time trend}
\usage{
\method{tidy}{disbayes_trend}(x, startyear = 0, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{disbayes}}}

\item{startyear}{Constant to add to the year variable in the result (e.g. to convert years 1-100 to years 1918-2018).}

\item{...}{Other arguments (currently unused)}
}
\description{
Tidy output for a disbayes model with a time trend
}
